Replace Netlify checkout function with Node API endpoint
