--- netlify.toml
@@ -1,14 +1,0 @@
-[build]
-  functions = "netlify/functions"
-
-[[redirects]]
-  from = "/api/stripe/create-checkout-session"
-  to = "/.netlify/functions/create-checkout-session"
-  status = 200
-  force = true
-
-[[redirects]]
-  from = "/api/create-checkout-session"
-  to = "/.netlify/functions/create-checkout-session"
-  status = 200
-  force = true
