+++ stripe-checkout-server.mjs
@@ -1,131 +1,61 @@
 #!/usr/bin/env node
 import { createServer } from 'node:http';
 import { createHmac, timingSafeEqual } from 'node:crypto';
-import fs from 'node:fs/promises';
 
 const port = Number(process.env.PORT || 4242);
 const stripeSecretKey = process.env.STRIPE_SECRET_KEY || '';
 const webhookSigningSecret = process.env.STRIPE_WEBHOOK_SECRET || '';
-const storePath = process.env.STRIPE_STORE_PATH || 'stripe-runtime-store.json';
 const allowedOrigin = process.env.STRIPE_ALLOWED_ORIGIN || '*';
 
-const defaultSuccessUrl =
-  process.env.STRIPE_SUCCESS_URL ||
-  'https://dashboard.stripe.com/workbench/blueprints/one-time-payment/checkout-chapter?confirmation-redirect=create-checkout-session';
-const defaultCancelUrl =
-  process.env.STRIPE_CANCEL_URL ||
-  'https://dashboard.stripe.com/workbench/blueprints/one-time-payment/checkout-chapter?confirmation-redirect=create-checkout-session';
+const defaultSuccessUrl = process.env.STRIPE_SUCCESS_URL || 'https://maybenot.com/success.html';
+const defaultCancelUrl = process.env.STRIPE_CANCEL_URL || 'https://maybenot.com/cart.html';
+
+function corsHeaders() {
+  return {
+    'Access-Control-Allow-Origin': allowedOrigin,
+    'Access-Control-Allow-Methods': 'POST, GET, OPTIONS',
+    'Access-Control-Allow-Headers': 'Content-Type, Stripe-Signature'
+  };
+}
 
 function jsonResponse(res, statusCode, payload) {
   res.writeHead(statusCode, {
     'Content-Type': 'application/json; charset=utf-8',
-    'Access-Control-Allow-Origin': allowedOrigin,
-    'Access-Control-Allow-Methods': 'POST, GET, OPTIONS',
-    'Access-Control-Allow-Headers': 'Content-Type, Stripe-Signature'
+    ...corsHeaders()
   });
   res.end(`${JSON.stringify(payload)}\n`);
 }
 
-function flattenFormBody(value, keyPrefix = '') {
-  const entries = [];
+function sanitizeLineItems(body) {
+  const directItems = Array.isArray(body?.lineItems) ? body.lineItems : [];
+  const direct = directItems
+    .map((item) => ({
+      price: typeof item?.price === 'string' ? item.price.trim() : '',
+      quantity: Number.isFinite(Number(item?.quantity)) ? Math.max(1, Number(item.quantity)) : 1
+    }))
+    .filter((item) => item.price.startsWith('price_'));
 
-  if (Array.isArray(value)) {
-    value.forEach((item, index) => {
-      const childKey = `${keyPrefix}[${index}]`;
-      entries.push(...flattenFormBody(item, childKey));
-    });
-    return entries;
-  }
+  if (direct.length) return direct;
 
-  if (value && typeof value === 'object') {
-    Object.entries(value).forEach(([key, childValue]) => {
-      const childKey = keyPrefix ? `${keyPrefix}[${key}]` : key;
-      entries.push(...flattenFormBody(childValue, childKey));
-    });
-    return entries;
-  }
-
-  if (value === undefined || value === null) {
-    return entries;
-  }
-
-  entries.push([keyPrefix, String(value)]);
-  return entries;
+  const cartItems = Array.isArray(body?.cart) ? body.cart : [];
+  return cartItems
+    .map((item) => ({
+      price: typeof item?.stripePriceId === 'string' ? item.stripePriceId.trim() : '',
+      quantity: Number.isFinite(Number(item?.quantity)) ? Math.max(1, Number(item.quantity)) : 1
+    }))
+    .filter((item) => item.price.startsWith('price_'));
 }
 
 function toFormBody(params) {
-  const body = new URLSearchParams();
-  flattenFormBody(params).forEach(([key, val]) => body.append(key, val));
-  return body;
-}
-
-async function stripeRequest(path, { method = 'GET', params } = {}) {
-  if (!stripeSecretKey) {
-    throw new Error('Missing STRIPE_SECRET_KEY environment variable.');
-  }
-
-  const response = await fetch(`https://api.stripe.com${path}`, {
-    method,
-    headers: {
-      Authorization: `Bearer ${stripeSecretKey}`,
-      'Content-Type': 'application/x-www-form-urlencoded'
-    },
-    body: params ? toFormBody(params) : undefined
+  const form = new URLSearchParams();
+  params.line_items.forEach((item, index) => {
+    form.append(`line_items[${index}][price]`, item.price);
+    form.append(`line_items[${index}][quantity]`, String(item.quantity));
   });
-
-  const payload = await response.json();
-  if (!response.ok) {
-    const message = payload?.error?.message || `Stripe API request failed with status ${response.status}.`;
-    throw new Error(message);
-  }
-  return payload;
-}
-
-async function loadStore() {
-  try {
-    const raw = await fs.readFile(storePath, 'utf8');
-    const parsed = JSON.parse(raw);
-    return {
-      product_id: parsed.product_id || '',
-      default_price_id: parsed.default_price_id || '',
-      checkout_sessions: Array.isArray(parsed.checkout_sessions) ? parsed.checkout_sessions : [],
-      completed_sessions: Array.isArray(parsed.completed_sessions) ? parsed.completed_sessions : []
-    };
-  } catch {
-    return {
-      product_id: '',
-      default_price_id: '',
-      checkout_sessions: [],
-      completed_sessions: []
-    };
-  }
-}
-
-async function saveStore(store) {
-  await fs.writeFile(storePath, `${JSON.stringify(store, null, 2)}\n`, 'utf8');
-}
-
-async function ensureExampleProduct(store) {
-  if (store.product_id && store.default_price_id) {
-    return { productId: store.product_id, defaultPriceId: store.default_price_id };
-  }
-
-  const product = await stripeRequest('/v1/products', {
-    method: 'POST',
-    params: {
-      name: 'Example Product',
-      default_price_data: {
-        currency: 'usd',
-        unit_amount: 2000
-      }
-    }
-  });
-
-  store.product_id = product.id;
-  store.default_price_id = product.default_price;
-  await saveStore(store);
-
-  return { productId: product.id, defaultPriceId: product.default_price };
+  form.append('mode', params.mode);
+  form.append('success_url', params.success_url);
+  form.append('cancel_url', params.cancel_url);
+  return form;
 }
 
 function parseStripeSignatureHeader(headerValue = '') {
@@ -169,55 +99,56 @@ async function readBody(req) {
 }
 
 async function handleCreateCheckoutSession(req, res) {
+  if (!stripeSecretKey) {
+    return jsonResponse(res, 500, { error: 'Missing STRIPE_SECRET_KEY.' });
+  }
+
   const rawBody = await readBody(req);
-  let body = {};
+  let body;
   try {
     body = rawBody ? JSON.parse(rawBody) : {};
   } catch {
     return jsonResponse(res, 400, { error: 'Invalid JSON body.' });
   }
 
-  const store = await loadStore();
-  const { defaultPriceId } = await ensureExampleProduct(store);
-  const incomingLineItems = Array.isArray(body.lineItems) ? body.lineItems : [];
-  const normalizedLineItems = incomingLineItems
-    .map((item) => ({
-      price: typeof item?.price === 'string' ? item.price : '',
-      quantity: Number.isFinite(Number(item?.quantity)) ? Math.max(1, Number(item.quantity)) : 1
-    }))
-    .filter((item) => item.price.startsWith('price_'));
-  const lineItems = normalizedLineItems.length
-    ? normalizedLineItems
-    : [
-        {
-          price: defaultPriceId,
-          quantity: 1
-        }
-      ];
+  const lineItems = sanitizeLineItems(body);
+  if (!lineItems.length) {
+    return jsonResponse(res, 400, { error: 'Request must include lineItems or cart with valid Stripe price IDs.' });
+  }
 
-  const session = await stripeRequest('/v1/checkout/sessions', {
-    method: 'POST',
-    params: {
-      line_items: lineItems,
-      mode: 'payment',
-      success_url: body.successUrl || defaultSuccessUrl,
-      cancel_url: body.cancelUrl || defaultCancelUrl
+  const successUrl = typeof body.successUrl === 'string' && body.successUrl ? body.successUrl : defaultSuccessUrl;
+  const cancelUrl = typeof body.cancelUrl === 'string' && body.cancelUrl ? body.cancelUrl : defaultCancelUrl;
+
+  const payload = {
+    line_items: lineItems,
+    mode: 'payment',
+    success_url: successUrl,
+    cancel_url: cancelUrl
+  };
+
+  try {
+    const response = await fetch('https://api.stripe.com/v1/checkout/sessions', {
+      method: 'POST',
+      headers: {
+        Authorization: `Bearer ${stripeSecretKey}`,
+        'Content-Type': 'application/x-www-form-urlencoded'
+      },
+      body: toFormBody(payload)
+    });
+
+    const stripePayload = await response.json();
+    if (!response.ok) {
+      const message = stripePayload?.error?.message || 'Stripe session creation failed.';
+      return jsonResponse(res, response.status, { error: message });
     }
-  });
 
-  store.checkout_sessions.push({
-    id: session.id,
-    url: session.url,
-    created: new Date().toISOString()
-  });
-  await saveStore(store);
-
-  return jsonResponse(res, 200, {
-    id: session.id,
-    url: session.url,
-    product_id: store.product_id,
-    default_price_id: store.default_price_id
-  });
+    return jsonResponse(res, 200, {
+      id: stripePayload.id,
+      url: stripePayload.url
+    });
+  } catch (error) {
+    return jsonResponse(res, 500, { error: error.message || 'Unable to create checkout session.' });
+  }
 }
 
 async function handleStripeWebhook(req, res) {
@@ -228,36 +159,13 @@ async function handleStripeWebhook(req, res) {
     return jsonResponse(res, 400, { error: 'Webhook signature verification failed.' });
   }
 
-  let event;
-  try {
-    event = JSON.parse(rawBody);
-  } catch {
-    return jsonResponse(res, 400, { error: 'Invalid webhook payload.' });
-  }
-
-  if (event.type === 'checkout.session.completed') {
-    const store = await loadStore();
-    store.completed_sessions.push({
-      event_id: event.id,
-      session_id: event.data?.object?.id || '',
-      customer: event.data?.object?.customer || '',
-      payment_status: event.data?.object?.payment_status || '',
-      received: new Date().toISOString()
-    });
-    await saveStore(store);
-  }
-
   return jsonResponse(res, 200, { received: true });
 }
 
 const server = createServer(async (req, res) => {
   try {
     if (req.method === 'OPTIONS') {
-      res.writeHead(204, {
-        'Access-Control-Allow-Origin': allowedOrigin,
-        'Access-Control-Allow-Methods': 'POST, GET, OPTIONS',
-        'Access-Control-Allow-Headers': 'Content-Type, Stripe-Signature'
-      });
+      res.writeHead(204, corsHeaders());
       res.end();
       return;
     }
